Update dependencies and configurations for improved security and compatibility
In this commit, multiple adjustments are made to improve the project's security measures and ensure compatibility with newer software versions. The 'testcontainers' dependency version is updated to 1.19.0, improving compatibility and potential bug fixes.

The annotation for Redis configuration is changed from "@Profile" to "@ConditionalOnProperty", offering more flexible conditions for it to be active.
Unnecessary dependencies in AuthService.kt, such as MemberRoleRepository and PasswordEncoder, are also removed to maintain a clean codebase.

Changes are made to application.yaml for security improvements, such as emptying the 'client-secret' and 'client-id'. Addition of configuration for an SSL connection and changes in profiles' specific configurations are also introduced for better environment separation and security.

Lastly, the TestController is adjusted to not run on 'prod' and 'dev' profiles, preventing potential accidental exposure during production or development.