Refactor code for readability and efficiency
Changes have been made across multiple files with a focus on improving readability and efficiency. Kotlin's features like 'with' function and 'let' function are utilized for better handling of objects. Conditional statements are shortened by using built-in functions like 'let', 'also', 'takeIf'. Unnecessary variables are removed to make the code cleaner. Exceptions are caught in smarter ways to reduce the number of lines and improve readability in function, especially in 'getUserOfExpiredToken' function. In authentication functions, steps are consolidated in fluent style to improve efficiency and readability. Furthermore, Optional is replaced with Kotlin's nullable types for consistency throughout the codebase.